Working create and read Asset
