+++ HomesEngland.Gateway.Test/PostgresConnectionFactoryTests.cs
@@ -1,5 +1,6 @@
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using HomesEngland.Gateway;
+using HomesEngland.Gateway.Sql.Postgres;
 
 namespace HomesEngland.Gateway.Test
 {
