+++ HomesEngland.Gateway.Test/PostgresConnectionStringFormatterTests.cs
@@ -1,4 +1,5 @@
 ﻿using FluentAssertions;
+using HomesEngland.Gateway.Sql.Postgres;
 using NUnit.Framework;
 
 namespace HomesEngland.Gateway.Test
