+++ HomesEngland.Gateway/InMemory/InMemoryAssetReader.cs
