+++ HomesEngland/Gateway/Assets/IAssetReader.cs
@@ -6,4 +6,9 @@ namespace HomesEngland.Gateway.Assets
     {
 
     }
+
+    public interface IAssetCreator : IDatabaseEntityCreator<IAsset, int>
+    {
+
+    }
 }
