+++ HomesEngland/UseCase/GetAsset/Impl/GetAssetUseCase.cs
@@ -1,6 +1,5 @@
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using HomesEngland.Exception;
-using HomesEngland.Gateway;
 using HomesEngland.Gateway.Assets;
 using HomesEngland.UseCase.GetAsset.Models;
 using Infrastructure.Api.Exceptions;
