+++ Main/AssetRegister.cs
@@ -3,6 +3,7 @@ using DependencyInjection;
 using HomesEngland.Gateway;
 using HomesEngland.Gateway.Assets;
 using HomesEngland.Gateway.Migrations;
+using HomesEngland.Gateway.Sql.Postgres;
 using HomesEngland.UseCase.GetAsset;
 using HomesEngland.UseCase.GetAsset.Impl;
 using InMemoryAssetReader = HomesEngland.Gateway.InMemoryAssetReader;
